function generateTarget -> generate a random number from 0 - 9
function compareGuesses -> takes three parameters (humanGuess),
(ComputerGuess) y el (secretTarget).

then it determine the absolute value of the human and computer guess.

then it returns false if absolute value of the (humanGuess) is
 greater than (computerGuess), returns true in every other scenario.

function updateScore -> takes one string as a parameter (winner) and
it update the players score adding one value to the player who wom.

function advanceRound -> add a value to the (currentRoungNumber).